slalom-consulting-bridgectl/bridgectl_v2.5.9/requirements.txt
colorama>=0.4.6
docker>=7.1.0
kubernetes>=30.1.0
packaging>=24.1
pandas>=2.2.2
pillow>=10.4.0
prompt-toolkit>=3.0.8
psutil>=6.0.0
PyGithub>=2.4.0  #futuredev: maybe just use requests library
PyYAML>=6.0.2
questionary>=2.0.1
requests>=2.32.3
rich>=13.8.1
slack_sdk>=3.33.0
streamlit>=1.41.1
tabulate>=0.9.0
urllib3>=2.2.3

#internal build only

